refactor: 뷰모델로 인해 프리뷰 안보임 - Content 뷰 분리
